requirements.txt
blessed==1.20.0
bpython==0.25
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
curtsies==0.4.2
cwcwidth==0.1.9
greenlet==3.1.1
idna==3.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyxdg==0.28
requests==2.32.3
six==1.17.0
urllib3==2.3.0
wcwidth==0.2.13
